Updated auton, better auton start/ending
Added autonomousRun() and autonomousStop() that starts and stops
autonomous as a new task, stopping in a safer way than before.

Autonomous is basically just feed and shoot.
Note: we found a bug that the waitForFlywheelReady seems to pass through
the first time, so a delay of 3 seconds was used. That is also the
reason for the spamming use of motorSet(8, 0)'s.

Note: not fully tested.